// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** Represents `true` or `false` values. */
  BooleanType: { input: unknown; output: unknown; }
  CustomData: { input: unknown; output: unknown; }
  /** A ISO 8601 compliant date value */
  Date: { input: unknown; output: unknown; }
  /** A ISO 8601 compliant datetime value */
  DateTime: { input: unknown; output: unknown; }
  /** Represents signed double-precision fractional values as specified by [IEEE 754](http://en.wikipedia.org/wiki/IEEE_floating_point). */
  FloatType: { input: unknown; output: unknown; }
  /** Represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1. */
  IntType: { input: unknown; output: unknown; }
  ItemId: { input: unknown; output: unknown; }
  MetaTagAttributes: { input: unknown; output: unknown; }
  UploadId: { input: unknown; output: unknown; }
};
// ...
export type ArticleRecord = RecordInterface & {
  _createdAt: Scalars['DateTime']['output'];
  /** Editing URL */
  _editingUrl?: Maybe<Scalars['String']['output']>;
  _firstPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  _isValid: Scalars['BooleanType']['output'];
  _modelApiKey: Scalars['String']['output'];
  _publicationScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Generates SEO and Social card meta tags to be used in your frontend */
  _seoMetaTags: Array<Tag>;
  _status: ItemStatus;
  _unpublishingScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _updatedAt: Scalars['DateTime']['output'];
  author: AuthorRecord;
  createdat: Scalars['Date']['output'];
  id: Scalars['ItemId']['output'];
  link?: Maybe<Scalars['String']['output']>;
  slug: Scalars['String']['output'];
  text: Scalars['String']['output'];
  title: Scalars['String']['output'];
};
// ...
export type AuthorRecord = RecordInterface & {
  _createdAt: Scalars['DateTime']['output'];
  /** Editing URL */
  _editingUrl?: Maybe<Scalars['String']['output']>;
  _firstPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  _isValid: Scalars['BooleanType']['output'];
  _modelApiKey: Scalars['String']['output'];
  _publicationScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Generates SEO and Social card meta tags to be used in your frontend */
  _seoMetaTags: Array<Tag>;
  _status: ItemStatus;
  _unpublishingScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _updatedAt: Scalars['DateTime']['output'];
  id: Scalars['ItemId']['output'];
  name: Scalars['String']['output'];
};
// ...
export enum ItemStatus {
  Draft = 'draft',
  Published = 'published',
  Updated = 'updated'
}
// ...
export type RecordInterface = {
  _createdAt: Scalars['DateTime']['output'];
  /** Editing URL */
  _editingUrl?: Maybe<Scalars['String']['output']>;
  _firstPublishedAt?: Maybe<Scalars['DateTime']['output']>;
  _isValid: Scalars['BooleanType']['output'];
  _modelApiKey: Scalars['String']['output'];
  _publicationScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Generates SEO and Social card meta tags to be used in your frontend */
  _seoMetaTags: Array<Tag>;
  _status: ItemStatus;
  _unpublishingScheduledAt?: Maybe<Scalars['DateTime']['output']>;
  _updatedAt: Scalars['DateTime']['output'];
  id: Scalars['ItemId']['output'];
};
// ...
export type Tag = {
  attributes?: Maybe<Scalars['MetaTagAttributes']['output']>;
  content?: Maybe<Scalars['String']['output']>;
  tag: Scalars['String']['output'];
};
// ...
export type ArticleCountQueryVariables = Exact<{ [key: string]: never; }>;


export type ArticleCountQuery = { allArticles: Array<{ slug: string }>, _allArticlesMeta: { count: unknown } };

export type ArticleGetBySlugQueryVariables = Exact<{
  slug?: InputMaybe<Scalars['String']['input']>;
}>;


export type ArticleGetBySlugQuery = { article?: { slug: string, title: string, text: string, createdat: unknown, link?: string | null } | null };

export type ArticleGetFirstQueryVariables = Exact<{
  quantity?: InputMaybe<Scalars['IntType']['input']>;
}>;


export type ArticleGetFirstQuery = { allArticles: Array<{ slug: string, title: string, text: string, createdat: unknown, link?: string | null }> };

export type ArticleGetListQueryVariables = Exact<{ [key: string]: never; }>;


export type ArticleGetListQuery = { allArticles: Array<{ slug: string, title: string, text: string, createdat: unknown, link?: string | null }> };

export type ArticleGetListWithPaginationQueryVariables = Exact<{
  first?: InputMaybe<Scalars['IntType']['input']>;
  skip?: InputMaybe<Scalars['IntType']['input']>;
}>;


export type ArticleGetListWithPaginationQuery = { allArticles: Array<{ slug: string, title: string, text: string, createdat: unknown, link?: string | null }> };

export type ArticleGetSlugListQueryVariables = Exact<{ [key: string]: never; }>;


export type ArticleGetSlugListQuery = { allArticles: Array<{ slug: string }> };

export type ArticleItemFragment = { slug: string, title: string, text: string, createdat: unknown, link?: string | null };

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
export const ArticleItemFragmentDoc = new TypedDocumentString(`
    fragment ArticleItem on ArticleRecord {
  slug
  title
  text(markdown: true)
  createdat
  link
}
    `, {"fragmentName":"ArticleItem"}) as unknown as TypedDocumentString<ArticleItemFragment, unknown>;
export const ArticleCountDocument = new TypedDocumentString(`
    query ArticleCount {
  allArticles {
    slug
  }
  _allArticlesMeta {
    count
  }
}
    `) as unknown as TypedDocumentString<ArticleCountQuery, ArticleCountQueryVariables>;
export const ArticleGetBySlugDocument = new TypedDocumentString(`
    query ArticleGetBySlug($slug: String) {
  article(filter: {slug: {eq: $slug}}) {
    ...ArticleItem
  }
}
    fragment ArticleItem on ArticleRecord {
  slug
  title
  text(markdown: true)
  createdat
  link
}`) as unknown as TypedDocumentString<ArticleGetBySlugQuery, ArticleGetBySlugQueryVariables>;
export const ArticleGetFirstDocument = new TypedDocumentString(`
    query ArticleGetFirst($quantity: IntType) {
  allArticles(first: $quantity, orderBy: createdat_DESC) {
    ...ArticleItem
  }
}
    fragment ArticleItem on ArticleRecord {
  slug
  title
  text(markdown: true)
  createdat
  link
}`) as unknown as TypedDocumentString<ArticleGetFirstQuery, ArticleGetFirstQueryVariables>;
export const ArticleGetListDocument = new TypedDocumentString(`
    query ArticleGetList {
  allArticles {
    ...ArticleItem
  }
}
    fragment ArticleItem on ArticleRecord {
  slug
  title
  text(markdown: true)
  createdat
  link
}`) as unknown as TypedDocumentString<ArticleGetListQuery, ArticleGetListQueryVariables>;
export const ArticleGetListWithPaginationDocument = new TypedDocumentString(`
    query ArticleGetListWithPagination($first: IntType, $skip: IntType) {
  allArticles(orderBy: createdat_DESC, first: $first, skip: $skip) {
    ...ArticleItem
  }
}
    fragment ArticleItem on ArticleRecord {
  slug
  title
  text(markdown: true)
  createdat
  link
}`) as unknown as TypedDocumentString<ArticleGetListWithPaginationQuery, ArticleGetListWithPaginationQueryVariables>;
export const ArticleGetSlugListDocument = new TypedDocumentString(`
    query ArticleGetSlugList {
  allArticles {
    slug
  }
}
    `) as unknown as TypedDocumentString<ArticleGetSlugListQuery, ArticleGetSlugListQueryVariables>;